+++ rule/errors.go
@@ -1,0 +1,28 @@
+// Copyright (C) 2016-2017 vLife Systems Ltd <http://vlifesystems.com>
+// Licensed under an MIT licence.  Please see LICENSE.md for details.
+
+package rule
+
+type InvalidRuleError struct {
+	Rule Rule
+}
+
+type IncompatibleTypesRuleError struct {
+	Rule Rule
+}
+
+type InvalidExprError struct {
+	Expr string
+}
+
+func (e InvalidRuleError) Error() string {
+	return "invalid rule: " + e.Rule.String()
+}
+
+func (e IncompatibleTypesRuleError) Error() string {
+	return "incompatible types in rule: " + e.Rule.String()
+}
+
+func (e InvalidExprError) Error() string {
+	return "invalid expression in rule: " + e.Expr
+}