+++ rule/rule_test.go
@@ -11,26 +11,6 @@ import (
 	"testing"
 )
 
-func TestInvalidRuleErrorError(t *testing.T) {
-	r := NewTrue()
-	err := InvalidRuleError{r}
-	want := "invalid rule: true()"
-	got := err.Error()
-	if got != want {
-		t.Errorf("Error() got: %s, want: %s", got, want)
-	}
-}
-
-func TestIncompatibleTypesRuleErrorError(t *testing.T) {
-	r := NewTrue()
-	err := IncompatibleTypesRuleError{r}
-	want := "incompatible types in rule: true()"
-	got := err.Error()
-	if got != want {
-		t.Errorf("Error() got: %s, want: %s", got, want)
-	}
-}
-
 func TestSort(t *testing.T) {
 	in := []Rule{
 		NewEQFV("band", dlit.MustNew("b")),